+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/FsOrphanFileFinderExecutor.java
@@ -27,10 +27,9 @@ import stroom.task.api.ThreadPoolImpl;
 import stroom.task.shared.ThreadPool;
 import stroom.util.io.FileUtil;
 import stroom.util.io.StreamUtil;
+import stroom.util.logging.LambdaLogger;
+import stroom.util.logging.LambdaLoggerFactory;
 import stroom.util.logging.LogExecutionTime;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -48,30 +47,29 @@ import javax.inject.Provider;
 /**
  * Task to clean the stream store.
  */
-class FsCleanExecutor {
+class FsOrphanFileFinderExecutor {
 
-    private static final String DELETE_OUT = "delete.out";
+    private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(FsOrphanFileFinderExecutor.class);
+    private static final String ORPHAN_FILE_LIST = "orphan_files.out";
+    public static final String TASK_NAME = "Orphan File Finder";
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(FsCleanExecutor.class);
     private final FsVolumeService volumeService;
     private final Duration oldAge;
-    private final boolean deleteOut;
-
-    private final Provider<FsCleanSubTaskHandler> fsCleanSubTaskHandlerProvider;
+    private final Provider<FsOrphanFileFinder> orphanFileFinderProvider;
     private final ExecutorProvider executorProvider;
     private final TaskContextFactory taskContextFactory;
     private final TaskContext parentContext;
     private final DataStoreServiceConfig config;
 
     @Inject
-    FsCleanExecutor(final FsVolumeService volumeService,
-                    final Provider<FsCleanSubTaskHandler> fsCleanSubTaskHandlerProvider,
-                    final ExecutorProvider executorProvider,
-                    final TaskContextFactory taskContextFactory,
-                    final TaskContext parentContext,
-                    final DataStoreServiceConfig config) {
+    FsOrphanFileFinderExecutor(final FsVolumeService volumeService,
+                               final Provider<FsOrphanFileFinder> orphanFileFinderProvider,
+                               final ExecutorProvider executorProvider,
+                               final TaskContextFactory taskContextFactory,
+                               final TaskContext parentContext,
+                               final DataStoreServiceConfig config) {
         this.volumeService = volumeService;
-        this.fsCleanSubTaskHandlerProvider = fsCleanSubTaskHandlerProvider;
+        this.orphanFileFinderProvider = orphanFileFinderProvider;
         this.executorProvider = executorProvider;
         this.taskContextFactory = taskContextFactory;
         this.parentContext = parentContext;
@@ -83,18 +81,18 @@ class FsCleanExecutor {
             age = Duration.ofDays(1);
         }
         this.oldAge = age;
-        this.deleteOut = config.isFileSystemCleanDeleteOut();
     }
 
-    public void clean() {
-        parentContext.info(() -> "Starting file system clean task. oldAge = " + oldAge);
+    public void scan() {
+        parentContext.info(() -> "Starting orphan file finder task. oldAge = " + oldAge);
+        final long oldestDirTime = System.currentTimeMillis() - oldAge.toMillis();
 
         final LogExecutionTime logExecutionTime = LogExecutionTime.start();
 
         final List<FsVolume> volumeList = volumeService.find(FindFsVolumeCriteria.matchAll()).getValues();
         if (volumeList != null && volumeList.size() > 0) {
             // Add to the task steps remaining.
-            final ThreadPool threadPool = new ThreadPoolImpl("File System Clean#",
+            final ThreadPool threadPool = new ThreadPoolImpl(TASK_NAME + "#",
                     1,
                     1,
                     config.getFileSystemCleanBatchSize(),
@@ -106,9 +104,9 @@ class FsCleanExecutor {
             for (final FsVolume volume : volumeList) {
                 if (VolumeUseStatus.ACTIVE.equals(volume.getStatus())) {
                     final Runnable runnable = taskContextFactory.childContext(parentContext,
-                            "Cleaning: " + volume.getPath(),
+                            "Checking: " + volume.getPath(),
                             taskContext ->
-                                    cleanVolume(taskContext, volume));
+                                    scanVolume(taskContext, volume, oldestDirTime));
                     final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable,
                             executor);
                     completableFutures[i++] = completableFuture;
@@ -117,36 +115,33 @@ class FsCleanExecutor {
             CompletableFuture.allOf(completableFutures).join();
         }
 
-        parentContext.info(() -> "start() - Completed file system clean task in " + logExecutionTime);
+        parentContext.info(() -> "start() - Completed orphan file finder in " + logExecutionTime);
     }
 
-    private void cleanVolume(final TaskContext taskContext, final FsVolume volume) {
-        final FsCleanProgress taskProgress = new FsCleanProgress();
-        if (deleteOut) {
-            final Path dir = Paths.get(volume.getPath());
-            if (!Files.isDirectory(dir)) {
-                LOGGER.error("Directory for file delete list does not exist '" + FileUtil.getCanonicalPath(dir) + "'");
-            } else {
-                final Path deleteListFile = dir.resolve(DELETE_OUT);
-                try {
-                    try (final PrintWriter printWriter = new PrintWriter(Files.newBufferedWriter(deleteListFile,
-                            StreamUtil.DEFAULT_CHARSET))) {
-                        final Consumer<List<String>> deleteListConsumer = list -> {
-                            synchronized (printWriter) {
-                                list.forEach(printWriter::println);
-                            }
-                        };
-
-                        final FsCleanSubTask subTask = new FsCleanSubTask(taskProgress, volume, "", "", oldAge, false);
-                        fsCleanSubTaskHandlerProvider.get().exec(taskContext, subTask, deleteListConsumer);
-                    }
-                } catch (final IOException e) {
-                    LOGGER.error("exec() - Error writing " + DELETE_OUT, e);
-                }
-            }
+    private void scanVolume(final TaskContext taskContext, final FsVolume volume, final long oldestDirTime) {
+        final Path dir = Paths.get(volume.getPath());
+        if (!Files.isDirectory(dir)) {
+            LOGGER.error(() -> "Directory for file delete list does not exist '" +
+                    FileUtil.getCanonicalPath(dir) +
+                    "'");
         } else {
-            final FsCleanSubTask subTask = new FsCleanSubTask(taskProgress, volume, "", "", oldAge, true);
-            fsCleanSubTaskHandlerProvider.get().exec(taskContext, subTask, null);
+            final Path orphanFileList = dir.resolve(ORPHAN_FILE_LIST);
+            try {
+                try (final PrintWriter printWriter = new PrintWriter(Files.newBufferedWriter(orphanFileList,
+                        StreamUtil.DEFAULT_CHARSET))) {
+                    final Consumer<Path> deleteConsumer = file -> {
+                        LOGGER.debug(() -> "Unexpected file in store: " +
+                                FileUtil.getCanonicalPath(file));
+                        synchronized (printWriter) {
+                            printWriter.println(FileUtil.getCanonicalPath(file));
+                        }
+                    };
+
+                    orphanFileFinderProvider.get().scanVolumePath(volume, deleteConsumer, oldestDirTime, taskContext);
+                }
+            } catch (final IOException e) {
+                LOGGER.error(() -> "exec() - Error writing " + ORPHAN_FILE_LIST, e);
+            }
         }
     }
 }